LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flex-basis-013-expected.txt


FAIL .container 1 assert_equals:
<div class="container row"><div data-expected-height="50" data-expected-width="50" class="item" style="flex-basis: auto;"><canvas height="5" width="5"></canvas></div></div>
width expected 50 but got 5
FAIL .container 2 assert_equals:
<div class="container row"><div data-expected-height="50" data-expected-width="50" class="item" style="flex-basis: content;"><canvas height="5" width="5"></canvas></div></div>
width expected 50 but got 5
FAIL .container 3 assert_equals:
<div class="container row"><div data-expected-height="50" data-expected-width="50" class="item" style="flex-basis: min-content;"><canvas height="5" width="5"></canvas></div></div>
width expected 50 but got 5
FAIL .container 4 assert_equals:
<div class="container row"><div data-expected-height="50" data-expected-width="50" class="item" style="flex-basis: fit-content;"><canvas height="5" width="5"></canvas></div></div>
width expected 50 but got 5
FAIL .container 5 assert_equals:
<div class="container row"><div data-expected-height="50" data-expected-width="50" class="item" style="flex-basis: max-content;"><canvas height="5" width="5"></canvas></div></div>
width expected 50 but got 5
PASS .container 6
PASS .container 7
PASS .container 8
PASS .container 9
PASS .container 10
PASS .container 11
PASS .container 12
PASS .container 13
PASS .container 14
PASS .container 15
PASS .container 16
PASS .container 17
PASS .container 18
PASS .container 19
PASS .container 20
